DS-217 build creates separate auth script
The build script now generates two scripts, one for starting the
container and a new one for authenticating the clerver account. The
start script is now named `start-irods` and the auth script is named
`auth-clerver`.  The start script is no longer written to standard
output. The build script now accepts an optional argument to
provide the path to the directory where the scripts will be
deposited.